clean up macro examples and README.md, remove run.dart script
